Renamed the validate.py file to standardise.py
