+++ nhs_number/__init__.py
@@ -26,5 +26,5 @@ from nhs_number.constants import (
 from nhs_number.constants import FULL_RANGE
 from nhs_number.details import NhsNumber
 from nhs_number.generate import generate
-from nhs_number.normalise import standardise_format
+from nhs_number.standardise import standardise_format
 from nhs_number.validate import is_valid, calculate_checksum
